app/requirements.txt
# Supernote OCR Enhancer Dependencies

# Supernote .note file parsing
supernotelib>=0.6.4

# HTTP client for OCR API
requests>=2.31.0

# FastAPI for health endpoint
fastapi>=0.109.0
uvicorn>=0.27.0

# Image handling
pillow>=10.0.0

# PDF generation with embedded OCR
reportlab>=4.0.0

# Data validation
pydantic>=2.0.0
